+++ src/app/search/_components/placeholders/SearchPlaceholder.tsx
@@ -1,27 +1,9 @@
 import KeywordList from "@/app/search/_components/placeholders/KeywordList";
+import SearchPlaceholderTitle from "@/app/search/_components/placeholders/SearchPlaceholderTitle";
 import useSearchTabStore from "@/stores/useTabStore";
 
-const MESSAGE_MAP = [
-  "아쉽게도 검색 결과가 없어요.",
-  "영화 검색 결과가 없어요.",
-  "톡 검색 결과가 없어요.",
-];
 const TAB_MAP = ["", "톡", "영화"];
 
-function SearchPlaceholderTitle({ tabIndex }: { tabIndex: number }) {
-  return (
-    <div className="flex flex-col items-center justify-center gap-2 py-[60px] text-center Tablet:gap-3 Tablet:py-[80px] Tablet:pb-0 Laptop:py-0  Desktop:gap-7">
-      <h2 className="text-Gray_Orange Text-m-Bold Tablet:Text-xl-Bold">
-        {MESSAGE_MAP[tabIndex]}
-      </h2>
-      <p className="text-L_Gray Text-s-Medium Tablet:Text-m-Medium">
-        단어의 철자가 정확한지 확인하거나
-        <br />
-        다른 단어로 다시 검색해보세요.
-      </p>
-    </div>
-  );
-}
 export default function SearchPlaceholder({
   relatedKeywords,
   isAlone,
@@ -35,6 +17,7 @@ export default function SearchPlaceholder({
     : activeSearchTab.includes("톡")
       ? 2
       : 0;
+
   if (activeSearchTab === "전체" && isAlone) {
     return (
       <div className="text-Gray Text-m-Bold Tablet:Text-l-Bold">
@@ -51,9 +34,7 @@ export default function SearchPlaceholder({
         <h2 className="text-Silver Text-s-Bold Tablet:Text-l-Bold">
           추천 검색어
         </h2>
-        {/* <Suspense fallback={<KeywordsSkeleton />}> */}
         <KeywordList relatedKeywords={relatedKeywords} />
-        {/* </Suspense> */}
       </div>
     </div>
   );
